Update patch set 1
Patch Set 1:

Tony,

PolicyEngine.py has a new param basic_client_auth in constructor to indicate whether to use or not to use the "Basic " for client-auth.  The removal of the whitespaces is a side effect of my changes to the file that is(was) not owned by dcae (yet).

Patch-set: 1
Reviewer: Gerrit User 1949 <1949@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0