Update patch set 1
Patch Set 1:

(1 comment)

nothing jumps out at me in the python code saying "this is wrong". It would be worth getting another pair of eyeballs on the python code.

One minor comment in one of the shell scripts. Please comment on that. If we get another set of eyeballs to do a +1 on the code, I'll approve it.

Patch-set: 1
Reviewer: Gerrit User 105 <105@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Verified=0